+++ components/detail/Bookmark.tsx
